requirements.txt
aniso8601==7.0.0
asgiref==3.2.7
Django==3.0.5
django-cors-headers==3.2.1
django-graphql-jwt==0.3.1
environs==7.3.1
graphene==2.1.8
graphene-django==2.9.1
graphql-core==2.3.1
graphql-relay==2.0.1
marshmallow==3.5.1
promise==2.3
PyJWT==1.7.1
python-dotenv==0.12.0
pytz==2019.3
Rx==1.6.1
singledispatch==3.4.0.3
six==1.14.0
sqlparse==0.3.1
